Another fix to windows compilation

+++ library/LibUtilities/LinearAlgebra/SparseDiagBlkMatrix.cpp
@@ -318,7 +318,7 @@ namespace Nektar
             typename SparseStorageType::const_iterator entry = m_submatrix[i]->begin();
             for (; entry != m_submatrix[i]->end(); ++entry)
             {
-                bandwidth = (std::max)(bandwidth, (IndexType) 2*abs(entry->first.first - entry->first.second)+1);
+                bandwidth = (std::max)(bandwidth, (IndexType) 2*abs( (int) entry->first.first - entry->first.second)+1);
             }
         }
         return bandwidth;
@@ -331,7 +331,7 @@ namespace Nektar
         typename SparseStorageType::const_iterator entry = m_submatrix[i]->begin();
         for (; entry != m_submatrix[i]->end(); ++entry)
         {
-            bandwidth = (std::max)(bandwidth, (IndexType) 2*abs(entry->first.first - entry->first.second)+1);
+            bandwidth = (std::max)(bandwidth, (IndexType) 2*abs( (int) entry->first.first - entry->first.second)+1);
         }
         return bandwidth;
     }
